Fix assert_poly_expansion for k > 1 (#275)
* Fix assert_poly_expansion for k > 1

The polynomial expansion was using a flattened tensor product when a normal kronecker product should have been used. This fixes that error.

As a consequence, it is now also possible to create a matrix from a
vector of vectors. This was implemented to make it easier to create the test cases for k > 1.