Fix commands using --features bulletproofs at workspace root
Ugh. Cargo saw "dep:bulletproofs" in the feature table, so it avoided
including the default "bulletproofs" feature for the optional
dependency. But whatever resolution cargo has at the workspace level for
passing features to members, it doesn't account for this and thinks it
_should_ try to pass "bulletproofs" to sunscreen_runtime. So, just
expose it as a feature to fix.