Fix linked proof commitments
The current graph traversal logic is depth first. Thus on the ZKP side,
the multiplier gates corresponding to the inputs will not be contiguous
as they are allocated into the prover's `Secret::{a_L,a_R}`.
In fact, determining the location of the input gates is not generally
feasible (any multiplications performed with said inputs are going to be
evaluated during the graph traversal as soon as those inputs are ..
inputted). If we instead change the graph traversal to be breadth first,
then we will allocate all the inputs first, in one contiguous block, and
_then_ we handle any downstream multiplications in the circuit.

Furthermore, since we're using a `VecDeque` to deque ready nodes, we may
as well pop them from the front. This allows for a more intuitive
traversal of the inputs in order, so we don't need to handle any
reversal or odd pending final multiplier case when constructing the
shared generators.

The tests pass, so it seems we didn't have any correctness assumptions
on this traversal.