Remove the ability to decrypt existing messages
As implemented, this would have almost never been used correctly. The
user would have had to ensure that the existing plaintext message had
the same plaintext polynomial encoding as the provided ciphertext. This
would generally only be true if both were fresh encryptions, in which
case this method is not even necessary.