Go over `User` before `Chain`
Now that the verication builder is in use, and the strategy is to call the same methods as the prover, it would be really confusing to read the verification part first. Since the user methods reference the `&Chain` type, also add an initial section where we define the two types.